Include some initial syntax definitions and help
hyperhelp requires a help "index" file that tells it all of the
appropriate meta information for the help in any package that might
support it. The first syntax added is basically just a simple wrapper
around the JSON syntax that associates a different top level scope to
the index files (which all have a fixed name).

The second syntax is an initial simple syntax for providing some of
the syntax elements that will be needed for upcoming development, such
as being able to detect when the cursor is on a link or where anchors
are.

In order to facilitate the testing of the syntax, a simple help file
has also been included. Currently the syntax has to be set manually
via the Sublime console, as there is no glue code yet.