Include support code for working with a help view
This includes three key pieces of code that we will need to use in
order to work with help.

First is a function that will create a brand new view in a given
window for use as help. This will set the tab text to an appropriate
bit of text, make sure that the buffer is a scratch buffer, and have
the appropriate syntax set.

Second is a function that will scan a given window in order to find a
previously created help view and return it back, so that its contents
can be modified.

Lastly is a function that uses the other two functions to display a
help string in a help view, either creating a new one if it does not
exist or reusing the existing one if it does.