Include some simple core code support
We now have a simple log function to use that will generate output to
the console (and also optionally to the status line and/or an error
dialog), ensuring that the message is prefixed so that people know
where it came from.

Also included is a function that takes a base syntax file name (for
example "HyperHelp.sublime-syntax") and returns back the full package
specification.