Include a simple test command
Include a simple test command that can be invoked to see if everything
we have so far works as expected. Invoking the command in a window
will open the sample help text file.

It is also possible to modify the source help file and run the command
again to see the help text change, in order to test not only the
initial help view creation but also the code to find an existing view.