Implement a new method for displaying help
This new method takes care of all of the details needed to get a help
file displayed in a view.

In particular any existing help view has its contents replaced with
the new help, or a new help view will be created first. In both cases
the help view is focused and the view specific settings that track
what is being displayed are updated.

In addition, these settings are now checked so that if the help file
being asked for is already being displayed, nothing else happens since
everything is already set up.

The test command now uses this new method to display the sample help
text.