Include post processing of loaded help files
Whenever a new set of help text is loaded into the window, we now do
some post processing on the help text.

Anything that is a link has a region applied that draws an underline
on it, allowing for better visual identification of what it is.

Also the hidden anchors are now processed. For each one we capture
their location in the buffer and modify them so that they no longer
have the syntax that marks them as hidden anchors.

Once processed, the list of anchor text and their locations in the
buffer is added as a view setting as a record of where they were for
later navigation.