Include a command to reload the current help file
In order to facilitate testing, this command will allow us to easily
reload the help file being displayed in the help view, so that help
changes are easier to visualize.

This works by altering the setting that indicates what file is
currently being displayed and then trying to load the existing help
file, which will cause it to be displayed again.

The setting change is necessary since the help loader is smart enough
to not reload the existing help file.