Include a context handler event listener
We want to bind a key to our new command, but there is a potential
issue. The key binding should only be active in a hyperhelp help view,
which is most easily distinguished by the scope of the text that it
contains.

An issue with this is that if we have a help file open for editing,
the scope will match, which is not what we want.

This adds a new custom context that will allow us to determine if a
help view is being authored or not so we can act accordingly.

This is technically part of authoring tools, but it is useful here for
testing, and can be moved later.