Include a key binding for reloading help
This finishes the operation of adding a way to reload the current help
file by actually providing a key binding that will invoke the command.

This uses two contexts; one to make sure that this is actually a help
file and another to make sure that the help file is not being authored
using our new context, so that it only triggers in the normal help
view.