Include new view functionality for focusing text
This adds in a new method named focus_on() to the view handling that
ensures that a given position (a point or a region) is visible by
focusing that part of the view and changing the selection.

This will be used in our navigation code to jump the help buffer to
the appropriate location.