Store all anchors in the view settings
In order to allow navigation between anchors more easily, this
modifies the existing post processing for anchors so that after all of
the hidden anchors are processed, the list of regular anchors is also
fetched and merged into the list.

The list is sorted so that the anchors appear in document order.

This obviates the need to try and determine at navigation time which
of the two sets of anchors has a position closest to where the cursor
is currently located.

Note that this may be redacted away by a future update since that sort
of logic will be needed to navigate between links and anchors, but
at the same time it seems like a good idea to have a definitive list
of anchors in the view, even if they can't currently be seen.