Include a command to navigate in the help
The beginnings of the help navigation command are now in place. This
command will take an argument that tells it where it should navigate
to and then perform the navigation.

The first navigation type included is the ability to jump through the
document landing on the next or previous hidden anchor.