Enhance the navigation command to follow links
The existing hyper_help_navigate command now supports an extra
navigation style that tells it to follow a link.

If this command is executed while the cursor is on a link, the list of
anchors in the current view is checked to see if the text of the link
is the same as an anchor.

When this is the case, the cursor is jumped to the appropriate
location and the view is focused on that position.

An error will be logged (and shown in the status line) if the text of
the link does not associate with an anchor in this document.