Support double clicking on links to follow them
This is done by listening for the drag select command with a "by"
argument that indicates the user is selecting by words.

If that is the case and the location that user double clicked on is a
link in a hyperhelp document, we manually invoke the command to
follow the link and block execution of the selection command.