Fix key bindings that were just added
The key bindings for following a link were set to always trigger
inside of a help file, which means that they block editing help text.

To fix this, the binding now uses our context to ensure that it does
not trigger in a help file that is being authored.

Just for an extra speed boost, the binding also now has a more dialed
in scope that only triggers the binding when the cursor is actively
inside of a link.