Enhance the help file syntax
This modifies the construct for key bindings so that instead of
disallowing all whitespace it instead specifically only disallows
spaces, so that tabs are allowed.

This allows for key bindings to contain tab characters to provide
visual spacing without making the help grammar ambiguous. Other
constructs such as links and anchors already had support for this.

Additionally a new separator is provided, which takes a sequence of
four or more `=` or `-` characters and syntax highlights them to add
some visual distinction.