Allow for case insensitive links
A modification to the newly added command to follow links causes it to
use a case insensitive search when trying to find the appropriate
anchor, so that it is possible to link to headings without matching
the case of the heading text for example.

This is done in a language-agnostic way so that regardless of the
language the user is using, this should still work.