Amend the internal settings to disable tabs
In particular, we force the tab size to be a single character and
disable the translation of tabs to spaces (for help files only).

The idea is that using tabs is a bad idea for help since the user may
not use the same tab size as you. Additionally, a tab that is a single
character wide is visually indistinguishable from a space but non-
ambiguous regarding the help grammar.