Start the parsing of the overall help index info
This includes a sample help index file and some of the supporting code
for loading and validating the help information. This will provide all
of the runtime information about the help in any particular package
without having to do any extra processing.

The bulk of this code is in validating that the structure of the JSON
is what is expected, kicking out warning or error messages as
appropriate.