Add missing file directive to topic cache
The code that creates the cache of all topics neglected to include the
file that the topic came from, which makes it impossible to follow a
topic to the appropriate location.

This is something that I noticed last night after I'd already pushed
up the commits for the day, because that's usually when you notice a
glaring error.