Include validictory to do schema validation
This removes the crunky self validation that was being done at load
time on the help index object in favor of using a validator to ensure
that the format of everything is as we expect before we even start.

This is done by including a copy of validictory by James Turk:
    https://github.com/jamesturk/validictory

This version 1.1.2 very slightly modified to work as a module stored
inside of our own module. As of January 2018 the author is deprecating
this library, but for our purposes (something small that is easy to
bundle inside our package), this is good enough for now.

This change includes a complete schema for the help index as it
currently exists, which means that some of the other code is now
cleaner because it can assume things are as we expect them to be.

The downside to this particular library is that the error messages it
generates are in no way human readable most of the time since it
insists on including the entire data object along with the error
message.

At some future point the validation library could be modified to
provide more useful error messages or be replaced with something
different.