Finalize the initial help index loading code
With the new validation of the index file in place, we now handle the
table of contents in the help file properly.

In cases where the appropriate key does not exist in the index, or
does exist but contains no entries, a table of contents will be
automatically created that contains all of the topics in the help
file, alphabetically sorted.

Otherwise we scan through all of the entries in the help contents list
and create a new version of the list that contains the full topic
entries for all of the items in the list, recursively.

This means that the entries that are expressed as just topic strings
become a complete topic dictionary, while those that are topic
dictionaries already potentially get a caption key inserted based on
the underlying topic if they don't have one.

This expansion is performed recursively across any children that the
entry may have as well.