Rename the core file for incoming changes
This change brings a little bit of cleanup to the code base. In
particular, the core.py file is going to be the file that contains the
core hyperhelp routines (the API, basically) which will be built up
from external code in the other modules.

In order for that to be the case, the current core.py has it's name
changed to be common.py (common global routines) so that there are not
circular dependency references between the external code that needs
the code in core.py (e.g. the log function).