Move the existing API command to the new core
The API function for displaying a help file is now in the new core
file, with ancillary changes elsewhere to adjust for the new location
of the function.

For this change the code in the test function is reverted to
displaying the help file instead of the loaded help index to allow for
testing.