Include the ability to scan for help indexes
A new help index internal method is added which performs a scan for
all of the help indexes available to known packages, which allows us
to easily discover all of the available help.

This uses the Sublime mechanism for finding all matching resources to
locate all of the hyperhelp.json files, which is why we use a specific
name for our help indexes.

As defined this function will only load the indexes for packages that
don't already exist, so that if we were to detect that there was a
new package we could rescan for help and only load the new ones.