Include new user facing API functions for indexes
There are now two new user facing API functions for interacting with
help index files.

The first of these is help_index_list(), which wraps the core
functionality of loading and reloading a global set of packages, for
use by the main hyperhelp code.

This function internally stores a list of all of the known help files
and transparently demand loads the list the first time you ask for it,
then returns the same list again.

Parameters are provided to allow you to force a reload of all help
index files or just the index for a specific package.

The second function is reload_help_index(), which takes a list of
loaded help files (as would be returned by help_index_list()) and
forcefully reloads one or all help indexes.

This is similar to help_index_list() but it works with an explicit
help index list instead of an implicit one, which allows external code
such as authoring tools to retain their own list of indexes that is
independent of the one maintained internally by the help system.