Alter the scope for help files to be more specific
Previously the scope applied to help text was text.hyperhelp, but this
causes a bit of ambiguity because the index files use the scope
text.hyperhelp.index, and so you might accidentally be able to target
an index file when you meant to just be a help file.

Now the scope is text.hyperhelp.help, so that things that want to
target a help file specifically can do so without issues.