Refactor the location of the plugin_loaded() code
This pulls the plugin_loaded() module function that is responsible for
making links properly underlined again at module load time (for when
help views are persisted in the session file) out of the help specific
code and into the core code, where it's more visible.

At a later point this will become part of the bootstrap code that
packages need to include to ensure that hyperhelp is all set up.