Expose the help index loader in the API
There is now a wrapper around the internal code for verifying and
normalizing a help index into an object, in case external codqe needs
to do so.

The underlying version has been modified to collect the package name
directly out of the package resource that contains the index\ instead
of requiring you to provide it yourself to make this easier.