Include an entirely new external API
This fleshes out a large part of the needed API in a way that can be
used both internally by the core help code as well as externally by
other packages that might want/need to get at some internals, such as
help authoring tools.

As a result of the new API functions, the commands are a little
simpler than they used to be.

This is not the final stable form of the API, since some code
reorganization and refactoring still needs to happen.