Include a command to reload help indexes
Since some upcoming steps are going to involve having to create more
help files, this includes a new command for reloading a help index.

This command would only be active from within a help index file and
will attempt to reload the help index, as long as the file that's open
is stored in the packages folder (where it will be a package
resource).

If the file is an appropriate package resource but is for a package
that does not have help loaded for yet, the command will perform a
rescan of all help indexes in order to pick it up, since there is
currently no API for demand loading a package index that does not
already exist.