Implement a command to view a help topic
This is an application command since there needs to only ever be one
of them.

This takes a package and a help topic and displays it. The package is
optional and will be inferred from the currently displayed help file
if it's not given.

The topic will default to index.txt because an upcoming change will
add a common header to all help files which includes their file name,
and also index.txt is the default base filename for help in any given
package.

Currently if there is no current help view and no package is given,
the command does nothing. In the future this will result in the user
being prompted to select a package first.