Include the ability to view table of contents
A new command is added which allows you to view the table of contents
for a help package. This displays a quick panel that shows the table
of contents as defined in the help index (or the default if one was
not specified). Picking an item opens that help topic, possibly also
opening up a help view is one does not already exist.

If no package is provided, then the package is inferred from the help
view in the current window, if there is one. If that doesn't yield a
package either, then the user is prompted to select a package first.

If prompt is set to True (default is False), the package is always
prompted for, even if it could be inferred or was already provided.
This allows for bindings that always allow you to select from all
available help if desired.