Include a syntax element to recognize help files
For the purposes of editing, it's handy to have the appropriate syntax
set so that as you're entering the help you can see it highlighting as
it will within the displayed help (although anchors are too wide and
links are not visualized with underlines).

Since we want help files to be largely human readable, they should
have an extension that makes that clear, which is why the defaults use
.txt for this.

Since it would not be a good idea for our syntax to highlight all text
files, we instead have a new rule in the syntax which only matches the
first line and uniquely represents a hyperhelp help file.

This also includes the ability to include some key=value pairs, which
will be used in an upcoming change.