Include a parse-able header line in help files
In order to bring extra structure to help documents now that there is
an identifying structure at the start of each file (potentially), we
now attempt to parse that header line and, if successful, replace it
with a more human readable version.

Specifically if the header line appears, it will be replaced with a
line that puts the name of the file as a target on the left hand side,
the modification date on the right hand side and the title centered in
the document.

The title will be clipped if it's too long, the date can be formatted
in any way the user chooses through a setting, a separator line is
generated under the header to visually separate it from the rest of
the help content, and the whole thing is syntax highlighted.