Enhance index loading to force help files in
With the new functionality for being able to show any help topic, we
introduced the idea that the default topic was index.txt to attempt to
impose the structure that like a web page the root topic is index.txt.

This is just so that no matter what, a user can get at help in a
package by knowing there is always at least one known help document.

Since all files with a common header now always have a link to
themselves in their header, this change forces into the index a topic
entry that links the name of a file to itself so that this works.

When this happens a default topic caption is used, but the help author
can override that default by creating their own topic entry in the
help index.