After post processing anchors, jump to BOF
After post processing anchors, the buffer is left with the cursor
focusing on the last hidden anchor that was processed, which leaves
the cursor in an indeterminate location.

Although this is generally harmless because anything opening a help
file is going to jump the current position to a topic anyway, this
does not happen when you reload an open help file, which leaves the
cursor in an odd state.