Add disabling of hyperhelp_contents in some cases
In particular, for the specific case of invoking this command with the
prompt parameter set to False (don't force prompt), the command will
disable itself if you didn't give it a package and it can't infer one
from a current help view.

This basically means that you said "Can I see the current table of
contents please?" when there is no help file open and thus no TOC to
view. In this case the command should be disabled.

Arguably if you want to be prompted you would set the prompt argument
to True to make it happen. Additionally if you wanted to see the
table of contents of a specific package, you would specify it.

This change means that the newly added entry in the command palette
for viewing the table of contents will disable itself (and thus be
hidden from the command palette) if there is not an open help file.