Fix a bug in reloading package indexes
When reloading a package index, if an attempt is made to reload a help
index for a package that was not previously loaded, the code would
react by invoking the internal command to reload all index files so
that the new one would get picked up.

An issue with this is that the internal reload command generates a new
help list in this case, but the command for reloading in this case was
not capturing the resulting list and storing it, so the package was
still not known.

This is resolved by having the command invoke the version of the API
method that stores the global help list directly, which is smart
enough to ensure that it updates the internal list in this case.