Enhance separators in the help syntax
This improves the separators in the help syntax (the horizontal
sequence of four or more `-` or `=` characters) so that the list may
optionally begin with a `+` character and may optionally end with a
`+` or `|` character.

Additionally a vertical separator is included, which is mapped to the
`|` character. Since this character is already used for links, the
syntax for a separator says that such a separator must have a leading
or trailing tab character.

In order to make this work, links (and anchors) are not allowed to
have leading tab characters immediately inside of the link so that
there is no ambiguity.