Update the help to include info on separators
The help page now talks about separators and how they work (including
a multi column table example) and mentioning the restriction on links
and anchors regarding leading or trailing whitespace.

This also demonstrates how in the table of contents you can override
the caption used for topics, so that the main topics for the markup
provided include an indication that they're talking about markup but
in the TOC they don't.