Include auto pair keys for some help elements
This includes some preliminary key bindings to automatically pair some
of the special characters in our markup.

These are duplicates of default bindings, so they may need more
tweaking in the future and/or they may have more context than they
strictly need at this point.

In addition the binding for code blocks may need to be amended or
removed when we add support for code blocks.