Clarify TOC command in the command palette
Immediately after pressing Post on the blog post for yesterday where I
said I didn't know of a way of blocking the keyboard binding from
showing up in the command palette even if it doesn't apply, I
immediately remembered that in fact I *do* know a way.

In particular Sublime naively compares the argument dictionary and the
command name to see if a key binding would take effect for a command
in the palette. Since both the binding and the palette are running the
command with no arguments, they appear to be the same.

I say "naively" here because in fact specifying even an empty argument
dictionary for one versus the other is enough to stop it from doing
that, because no arguments (i.e. all defaults) is considered different
than {} (i.e. an empty object, so all defaults).

Doing that in the command palette or binding would be enough to solve
the problem, but it looks somewhat ugly. A better fix is to do what is
done here, which is to specifically say for the version of the command
that is in the command palette that it should not prompt, which is the
default but makes it clear what's going to happen.

The key binding doesn't need this because the context constrains it to
only work in a help file, where it doesn't need to be told not to
prompt because it will always find a package anyway.