Fix a regression in the link syntax
The change made last night to stop leading tab characters from being
allowed in link text (so that the same character can be used as a
vertical separator unambiguously) had an inadvertent error in that it
inherently required link text to be two characters long (with the
first not being a selection of things including a tab).

The effect of this is that links of a single character in length were
not allowed. While probably not a problem because such a link seems
somewhat useless, it should still work.