Focus navigation items not in the center of pane
Previously while navigating around in help, such as when skipping
through anchors with the Tab key or following a link, we made sure to
focus the link in the center of the pane always.

This change modifies that behaviour so that regular navigation just
focuses the view, moving it as minimally as possible but while making
sure there are still some context lines.

In contrast following a link causes the view to center on the given
location instead, so that in-file links always brings the target
location into the best focus.