Factor out code to prompt user for a package
In preparation for upcoming commands needing to prompt the user for a
help package the same way that the code for displaying the help table
of contents or list of help packages does, this code has been factored
out.

This new method takes a list of help packages, prompts the user, then
invokes one of two callbacks (if they are provided) to tell the caller
what package was selected.