Don't include Packages/ prefix in doc root
Now that we're getting into the authoring tools in more detail, it
has come to light that having the document root contain the Packages/
part of the resource prefix is problematic for creating local files.

This change makes the document root be inherently relative to the
Packages folder (that is, it starts with the name of the package
instead), so that it can be easily added where needed.