Implement an authoring command to create help
This implements a new authoring command to make creating new help
files easier. If there is a current help file open, it will attempt to
create a help file in that package, otherwise it prompts you for the
package to use.

After you enter the name of the help file to create, a new stub help
file is created with the default directory and name set as appropriate
to make saving the new help file easier.

In the future this might be better served by actually saving the file
to disk as well, so that saving just natively works without being
prompted for a name, although that would require it to be smarter
about file names.

Also the help index is not updated as it probably should be to include
information on the new file, since that would require more input.