Enhance the authoring command for new help
One last bit of business left over from last night is an enhancement
to the authoring command for creating a new help file.

The base version was simple in that it expected you to enter only a
filename, and the final folder had to exist previously, or the save
would go wrong.

This version fixes that by combining together the document root with
the filename input, then splitting the filename off and ensuring that
the resulting directory exists.