Factor out some common code for authoring portions
Some code is probably going to be common to the authoring parts of the
package, so to get started a new common file for the authoring  tools
has been factored out with a couple of entries.

One is a rename of existing code for reformatting a template, while
the other is something that will come in handy shortly.