Include command to update help header date
This introduces a new command that will check the current file to see
if it is a help source file with a header that contains a date field.

If this is true and the date in the header is not current (i.e.
today), the date in the header will be modified to be the current
date.