Automatically invoke new command on save
With the ability to update the date header in a file on demand, this
includes a new event handler that runs this command on save, so that
as you edit help files the information in them is always kept up to
date.

This also includes the core for being able to have custom settings
for the authoring tools so that this operation can be configured.

As a result of this, the resources for the authoring tools are now in
a separate folder than the ones for the core help system, to better
define how they will be split apart at the end of Devember.