Augment the help system menu and key bindings
This modifies the main menu so that there is a preferences section for
HyperHelp contained in it. The settings for HyperHelp are stored in
the global user preferences currently, so this only allows for
modification of the key bindings.

To make this work as expected, the regular non-platform key bindings
have been swapped for platform specific versions so that they can
easily be opened, even though (currently) they are always the same
keys for all platforms.