Sort the list of help files in the help index
Due to a mild brain fart, I went out of my way originally to ensure
that the dictionary that associates help files with their names was an
ordered dictionary so that we could traverse it in sorted order, but
forgot to actually sort the keys before inserting them.

This is now fixed.