Include a command to open an existing help file
This adds a new command in the authoring tools that allows you to
easily open a help file for editing.

Like most authoring tool commands so far, this takes as parameters a
package and a prompt argument and does the now-usual use the package
or prompt for one thing.

Once a package is selected (if needed), you see a list of files that
exist in that help file's index and choosing one opens it for editing.

This opens the file and forces the appropriate syntax onto the file,
since it might not open with the correct syntax if it doesn't have an
appropriate help header.

Once that is done, a set of settings that is controlled via the
settings file is applied to override the syntax specific settings, so
that some aspects that you want disabled while viewing help are
available while editing it.