Normalize line endings on loaded help files
While testing things out on Windows, I discovered that the
load_resource() method returns you the raw data in the file without
regard for the line end characters that are in use.

The command that is used to populate the help text inserts the text
you give it verbatim, which seems to bypass Sublime's internal line
normalization.

The result of this is that since internally Sublime seems to use a
Unix termination philosophy, if the source help file is in the Windows
or Mac encodings, the help is not interpreted properly. In particular
a Windows file will show all of the carriage return characters at the
end of every line, while a Mac file will be one long line with
embedded carriage returns.

In order to resolve this, the help file loader detects when this is
happening and normalizes the text back to the proper line termination.

This is potentially costly (for memory if not time) if the help file
is large, so it's probably best to try and somehow ensure that help
files always use a particular line encoding.  That's something that
needs to be researched a bit, however.