Retain syntax when saving new help the first time
A consequence of reusing the txt extension for help files (because
they are meant to be human readable as is) is that when you save one
for the first time and are prompted for a file name, the syntax gets
reverted back to plain text.

I'm not currently willing to go to a custom extension, so for now a
newly created help view gets assigned the same custom setting that is
used to ensure that help files are loaded with the correct syntax.

A new event handler for on_pre_save detects if that setting is present
and if so it re-applies the authoring settings (since the syntax gets
changed to plain text before the event triggers).

A downside to this is that the view tends to flicker for a second
as the syntax swaps in and out, which is sub-optimal. For the time
being it's good enough, though.