Enhance the commands for creating/editing help
The existing authoring commands have been extended so that they take
an optional file argument as well.

This works as one would expect; if you provide both, that file is
opened/created directly without prompt, if you specify a project you
only get prompted for a file, and if you only specify a file you get
prompted for a package.