Include a simple package scanner
For an upcoming change, it's important to be able to know the list of
packages that are currently installed in Sublime, including the ones
that ship with it directly.

The more robust way to determine that information is to perform a scan
for all sublime-package files in Installed Packages and the Packages
sub-folder of the Sublime install folder, plus any loose directories
that exist in the Packages folder itself. OverrideAudit does this, for
example.

The code here is slightly more naive about this in that it asks
Sublime for all of the resource files that it knows about (basically
every package file), and then collects the unique set of package
names from there.

This works, although it is less than ideal as it doesn't find packages
that are currently ignored. In our case that doesn't matter because we
rely on load_resource() and find_resource() to get at the indexes and
help files and load them, and those calls won't work for ignored
packages anyway.

Along with this change, the default key bindings for Windows and MacOS
have been synchronized with the Linux bindings so that it's easier to
test on other platforms.