Include the ability to add help to a package
This is the last of the initially planned authoring commands, which
prompts you for a package (unless you provide one explicitly) and then
sets up a help system inside that package.

This works by asking you for the document root to use, then creating a
stub hyperhelp.json and index.txt file inside of it in the package
folder, making sure not to clobber away any file that might exist.

If that works, you are prompted to open the new files.

That last part is potentially fragile in that Sublime doesn't
immediately notice when new resources are added to a package, and
doesn't get a chance to handle the system notifications that they're
there until the current command ends.

This is good enough for now, but in the future it would be better to
handle this differently. Possibly a background thread would be spawned
that periodically checks to see if the resource is available, and then
reloads the help indexes once it happens.