Allow help indexes to load from local paths
This is a modification to the load_help_index() API call that is used
to load, validate and normalize a hyperhelp.json index file so that
besides just a package resource specification, it can also load an
index from a local file.

In order for this to transparently work, the input file specification
has to be rooted in the sublime package path.

If the incoming file is a resource specification, it's loaded the way
it always has been, while if it's an absolute file path to a file in
the packages directory, it's loaded using regular file system access
and then the file name is converted to the appropriate resource
specification for later.

The returned object is thus identical either way, and still refers to
the appropriate resource should anything need to know.