Remove hacky timeout call on help index creation
Using the new ability to load a help index directly from a local file
the command for adding help to a new file can now remove the hacky
call to delay loading the indexes.

Instead, it can directly load the specific index file it just created
and patch it into the global help list. This is arguably better anyway
since it doesn't force the already loaded indexes to be loaded again
for no reason.

It seems from observation that you can load a resource if you know the
name of it, you just can't find it for a few milliseconds.

If that's true in all cases, the previous change was not needed and
indeed could have done what this is doing but with a resource
specification version of the name that was manually constructed.

The command for displaying the newly created help topic might fail if
this is not always the case, in which case the code for loading a
help file would have to be modified in a similar fashion.

For the time being this will be left as it is for time reasons.