Default topic caption for file now matches title
When the index loader is pulling in the index, it automatically adds a
topic entry for any help file that doesn't include an anchor that
points to itself, so that it's always possible to open the file.

Previously the caption used for the topic was a string that pointed
out that it was a default to remind you that you should fix it.

After working on some help I noticed that I was almost always creating
the topic to have a caption the same as the string that defines the
help file title.

This change alters the default caption in this case to just match the
file. This means that just creating an entry for the help file the
required title string is all you need to have the file in the TOC.

As a result, the snippet for adding a help file now skips the
redundant topic, since you mostly always want it to not be there.

Testing of this change also exposed a bug in validictory, which seems
to not validate it's own input while validating the input, as it
throws an exception while trying to validate a help file array that is
completely empty.

I haven't investigated this fully as of yet, so it's something to keep
in mind for the future.