Include an index command for easier help searching
This introduces a new core help command which shows a sorted list of
every topic that is known in the help system for the current package,
allowing you to quickly filter down to any particular topic if you
know it exists but not where.

This is functionally similar to the table of contents command and in
fact will do the same thing as that command if the help index does not
include an explicit table of contents.

At some point the system may provide some sort of free text search
functionality, although if that is the case it will be a ways down the
road.

Also this functionality could probably be rolled into the existing
contents command (with an appropriate argument to select), but for the
time being this is good enough, since a big refactor phase has to
happen anyway.